% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RLumModel-package.R
\docType{data}
\name{ExampleData.FittingTL}
\alias{ExampleData.FittingTL}
\alias{TL_fitting_data}
\title{Example data with TL curves extracted from a TL-SAR protocol (lab code BT1195)}
\format{A RLum.Analysis object containing measured TL curves.}
\description{
Example data with TL curves extracted from a TL-SAR protocol (lab code BT1195)
}
\section{Function version}{
 0.1.0
}
\examples{

data("ExampleData.FittingTL", envir = environment())

}
\author{
Johannes Friedrich, University of Bayreuth (Germany)
}
\seealso{
\code{\link[FME]{modCost}}
}
\keyword{datasets}

