% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RLumModel-package.R
\docType{data}
\name{ExampleData.ModelOutput}
\alias{ExampleData.ModelOutput}
\alias{model.output}
\title{A data set containing a TL curve simulated from Bailey 2001 (see fig. 1)}
\format{RLum.Analysis object}
\source{
\bold{model_LuminescenceSignals()}
}
\usage{
data(ExampleData.ModelOutput)
}
\description{
A data set containing a TL curve simulated from Bailey 2001 (see fig. 1)
}
\note{
This example has only one record (TL). The used sequence was
sequence <- list(IRR = c(temp = 20, dose = 10, DoseRate = 1),
                 TL = c(temp_begin = 20, temp_end = 400, heating_rate = 5))
}
\section{Version}{
 0.1
}
\examples{

data(ExampleData.ModelOutput,envir = environment())
plot_RLum.Analysis(model.output)

}
\references{
Bailey, R.M., 2001. Towards a general kinetic model for optically and thermally stimulated
luminescence of quartz. Radiation Measurements 33, 17-45.
}
\keyword{datasets}

