% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_concentrations.R
\name{plot_concentrations}
\alias{plot_concentrations}
\title{Plot electron/hole concentrations of a specific record.id for}
\usage{
plot_concentrations(object, record.id, plot.saturation = FALSE, ...)
}
\arguments{
\item{object}{\code{\linkS4class{RLum.Analysis}} (\bold{required}):  S4
object of class \code{RLum.Analysis}, e.g. the values of \code{\link{model_LuminescenceSignals}}.}

\item{record.id}{\code{\link{numeric}} (\bold{required}): id of the simulated record, which
is to plot. To see all record.ids use \code{\link[Luminescence]{structure_RLum}}, see examples.}

\item{plot.saturation}{\code{\link{logical}} (with default): plots the saturation of every
level from a specific model. Default: FALSE.}

\item{\dots}{further arguments and graphical parameters passed to
\code{\link{plot.default}} and \code{\linkS4class{RLum.Analysis}}.}
}
\value{
Returns multiple plots for the concentrations of electrones/holes.
}
\description{
The functions provides a plot of all changes in time of the electron respectively hole concentration
in electron traps, hole centres, in the condunction and valence band.
}
\details{
The function produces a multiple plot output and uses in main parts the Luminescence function
\code{\link[Luminescence]{plot_RLum.Analysis}}. A file output is recommended (e.g., \code{\link{pdf}}).
}
\section{Function version}{
 0.1.0
}
\examples{

##load data
data(ExampleData.ModelOutput, envir = environment())

##show structure
Luminescence::structure_RLum(model.output)

##plot all concentrations
plot_concentrations(object = model.output, record.id = 1)

##plot only specific energy-band-level (e.g. 110 degree celsius trap, "concentration level 1")
plot_concentrations(object = model.output,
                    record.id = 1,
                    subset = list(recordType = "concentration level 1"))

##plot every level on a single plot
plot_concentrations(object = model.output, record.id = 1, plot.single = TRUE)

}
\author{
Johannes Friedrich, University of Bayreuth (Germany),
}
\references{
Bailey, R.M., 2001. Towards a general kinetic model for optically and thermally stimulated
luminescence of quartz. Radiation Measurements 33, 17-45.

Bailey, R.M., 2002. Simulations of variability in the luminescence characteristics of natural
quartz and its implications for estimates of absorbed dose.
Radiation Protection Dosimetry 100, 33-38.

Bailey, R.M., 2004. Paper I-simulation of dose absorption in quartz over geological timescales
and it simplications for the precision and accuracy of optical dating.
Radiation Measurements 38, 299-310.

Pagonis, V., Chen, R., Wintle, A.G., 2007: Modelling thermal transfer in optically
stimulated luminescence of quartz. Journal of Physics D: Applied Physics 40, 998-1006.

Pagonis, V., Wintle, A.G., Chen, R., Wang, X.L., 2008. A theoretical model for a new dating protocol
for quartz based on thermally transferred OSL (TT-OSL).
Radiation Measurements 43, 704-708.
}
\seealso{
\code{\link{plot}}, \code{\link[Luminescence]{plot_RLum.Analysis}}
}

