% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_SEQ2R.R
\name{read_SEQ2R}
\alias{read_SEQ2R}
\title{Parse a Risoe SEQ-file to a sequence neccessary for simulating quartz luminescence}
\usage{
read_SEQ2R(file, lab.DoseRate = 1, txtProgressBar = TRUE)
}
\arguments{
\item{file}{\code{\link{character}} (\bold{required}): a *.seq file created by the Risoe Sequene Editor}

\item{lab.DoseRate}{\code{\link{character}} (with default): set the doserate of the radiation source
in the laboratory [Gy/s]. Default: 1 Gy/s}

\item{txtProgressBar}{\code{\link{logical}}: enables or disables the txtProgressBar}

\item{\dots}{further arguments and graphical parameters passed to
\code{\link{plot.default}}. See details for further information}
}
\value{
This function returns a list with the parsed *.seq file and the required steps for
\code{\link{model_LuminescenceSignals}}
}
\description{
A SEQ-file created by the Risoe Sequence Editor can be imported to simulate the sequence written
in the sequence editor.
}
\details{
\bold{Supported versions}\cr

Supppored and tested: version 4.36.\cr
}
\section{Function version}{
 0.1.0
}
\examples{

#so far no example available

}
\author{
Johannes Friedrich, University of Bayreuth (Germany),
}
\seealso{
\code{\link{plot}}, \code{\link{model_LuminescenceSignals}}, \code{\link{readLines}}
}

