% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RLumModel-package.R
\docType{package}
\name{RLumModel-package}
\alias{RLumModel-package}
\title{Solving Ordinary Differential Equations to Unterstand Luminescence}
\description{
A collection of function to simulate luminescence signals in the mineral quartz based on
published models.
}
\details{
\tabular{ll}{ Package: \tab RLumModel\cr Type: \tab Package\cr Version:
\tab 0.2.3 \cr Date: \tab 2017-11-22 \cr License: \tab GPL-3\cr }

**Authors**

\tabular{ll}{
Johannes Friedrich \tab University of Bayreuth, Germany \cr
Sebastian Kreutzer \tab IRAMAT-CRP2A, Universite Bordeaux Montaigne, France\cr
Christoph Schmidt \tab University of Bayreuth, Germany
}

**Supervisor**

Christoph Schmidt, University of Bayreuth, Germany\cr
Sebastian Kreutzer, IRAMAT-CRP2A, Universite Bordeaux Montaigne, France

**Support contact**

\email{developers@model.r-luminescence.de}

**Project source code repository**
-[https://github.com/R-Lum/RLumModel]

**Related package projects**
- [http://www.r-luminescence.de]
- [https://cran.r-project.org/package=Luminescence]
- [http://shiny.r-luminescence.de]
- [https://cran.r-project.org/package=RLumShiny]

**Package maintainer**

Johannes Friedrich, University of Bayreuth, Germany
\cr \email{johannes.friedrich@posteo.de}

**Acknowledgement**

 The work of Johannes Friedrich is gratefully supported by the DFG in framework of the project
 'Modelling quartz luminescence signal dynamics relevant for dating and dosimetry' (SCHM 305114-1).
}
\keyword{package}
