% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RLumModel-package.R
\docType{package}
\name{RLumModel-package}
\alias{RLumModel}
\alias{RLumModel-package}
\title{Solving Ordinary Differential Equations to Understand Luminescence}
\description{
\if{html}{
\figure{RLumModel_Logo.png}{options: width="50"}\cr
}
}
\details{
A collection of function to simulate luminescence signals in the mineral quartz based on
published models.

\strong{Authors}

\tabular{ll}{
Johannes Friedrich \tab University of Bayreuth, Germany \cr
Sebastian Kreutzer \tab Geography & Earth Sciences, Aberystwyth University, United Kingdom\cr
Christoph Schmidt \tab University of Bayreuth, Germany
}

\strong{Supervisor}

Christoph Schmidt, University of Bayreuth, Germany\cr
Sebastian Kreutzer,Geography & Earth Sciences, Aberystwyth University, United Kingdom
#'
\strong{Project source code repository}
\itemize{
\item https://github.com/R-Lum/RLumModel
}

\strong{Related package projects}
\itemize{
\item http://r-luminescence.de
\item https://CRAN.r-project.org/package=Luminescence
\item https://CRAN.r-project.org/package=RLumShiny
}

\strong{Package maintainer}

Johannes Friedrich, University of Bayreuth, Germany
\cr \email{johannes.friedrich@posteo.de}

\strong{Acknowledgement}

The work of Johannes Friedrich is gratefully supported by the DFG in framework of the project
'Modelling quartz luminescence signal dynamics relevant for dating and dosimetry' (SCHM 305114-1).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://CRAN.R-project.org/package=RLumModel}
  \item Report bugs at \url{https://github.com/R-Lum/RLumModel/issues}
}

}
\author{
\strong{Maintainer}: Johannes Friedrich \email{johannes.friedrich@posteo.de} (\href{https://orcid.org/0000-0002-0805-9547}{ORCID}) [translator]

Authors:
\itemize{
  \item Sebastian Kreutzer (\href{https://orcid.org/0000-0002-0734-2199}{ORCID}) [thesis advisor]
  \item Christoph Schmidt (\href{https://orcid.org/0000-0002-2309-3209}{ORCID}) [thesis advisor]
}

Other contributors:
\itemize{
  \item Marco Colombo (\href{https://orcid.org/0000-0001-6672-0623}{ORCID}) [contributor]
}

}
\keyword{package}
